Revert "chore: Update badges in README"
This reverts commit de22cd10c8e118dec7234b9ed9bed05b4ed2af28.

This needs to be reverted, because it doesn't pass CI even though
it passes locally for me. With this reverted, it will fail locally
again. This is very curious, because any differences in file
sorting sue to different shells should be accounted for by the bash
shebang in the script.