ci(build): Add runners selection (#1299)
* ci(build): Add runners selection

* ci(build): Use jq to produce JSOn without newlines

* ci(build): Remove superfluous cat from command

* temp(ci/opa): Use only Ubicloud runners

This ensures that the build succeeds with the set of Ubicloud
runners specified in the JSON file.

* Revert "temp(ci/opa): Use only Ubicloud runners"

This reverts commit 15ef0fb963f6202a1360b0fa44285d8cd5247651.

* ci(opensearch-dashboards): Use Ubicloud runners

* chore: Update actionlint config file

* chore: Update badges in README

* Revert "chore: Update badges in README"

This reverts commit de22cd10c8e118dec7234b9ed9bed05b4ed2af28.

This needs to be reverted, because it doesn't pass CI even though
it passes locally for me. With this reverted, it will fail locally
again. This is very curious, because any differences in file
sorting sue to different shells should be accounted for by the bash
shebang in the script.